+++ html/src/App.tsx
@@ -1,5 +1,22 @@
 import { AuthProvider, useAuth } from "./auth.tsx";
-import { RouterProvider } from "@tanstack/react-router";
+import { createRouter, RouterProvider } from "@tanstack/react-router";
+import { routeTree } from "./routeTree.gen.ts";
+
+// Set up a Router instance
+const router = createRouter({
+  routeTree,
+  defaultPreload: 'intent',
+  context: {
+    auth: undefined!, // This will be set after we wrap the app in an AuthProvider
+  },
+})
+
+// Register things for typesafety
+declare module '@tanstack/react-router' {
+  interface Register {
+    router: typeof router
+  }
+}
 
 function InnerApp() {
   const auth = useAuth()
