+++ html/src/main.tsx
@@ -1,26 +1,7 @@
 import ReactDOM from 'react-dom/client'
-import { createRouter } from '@tanstack/react-router'
-
-import { routeTree } from './routeTree.gen'
 
 import './index.css'
 import { App } from "./App.tsx";
-
-// Set up a Router instance
-const router = createRouter({
-  routeTree,
-  defaultPreload: 'intent',
-  context: {
-    auth: undefined!, // This will be set after we wrap the app in an AuthProvider
-  },
-})
-
-// Register things for typesafety
-declare module '@tanstack/react-router' {
-  interface Register {
-    router: typeof router
-  }
-}
 
 const rootElement = document.getElementById('root')!
